refactor: rename files of services
