+++ studio_app/services/email_service.py
