+++ studio_app/services/verification_service.py
@@ -1,4 +1,4 @@
-from ..services.sms import SmsService
+from .sms_service import SmsService
 from ..config import Config
 
 class SmsVerificationOfBooking:
